Add update for measurementGroup (still not solved how to deal with field updates that already have some data)

+++ featureMeasurement/src/main/java/cz/vvoleman/phr/featureMeasurement/data/datasource/room/field/NumericFieldDataSourceModel.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "numeric_field")
 data class NumericFieldDataSourceModel(
-    @PrimaryKey(autoGenerate = true) val id: Int? = null,
+    @PrimaryKey(autoGenerate = true) val id: Long? = null,
     val name: String,
     @ColumnInfo(name = "measurement_group_id") val measurementGroupId: Int,
     val unitGroupId: Int,
